requirements.txt
faiss-cpu==1.7.3
ffmpeg-python==0.2.0
ipyaggrid==0.3.2
keract==4.5.1
keras==2.11.0
numpy==1.21.6
opencv-python==4.7.0.68
pandas==1.3.5
Pillow==9.4.0
plotly==5.13.0
scipy==1.7.3
seaborn==0.12.2
tensorflow==2.11.0
tqdm==4.64.1
Wand==0.6.11